Move temp_set_env from environ.py to utils/common.py
temp_set_env is a general-purpose env helper that explicitly rejects
sglang-owned (SGLANG_*/SGL_*) keys, so it belongs alongside the other
env helpers in utils/common.py rather than in environ.py. Update all
callers accordingly.
