+++ python/sglang/test/kits/mmmu_vlm_kit.py
@@ -7,7 +7,7 @@ import tempfile
 from pathlib import Path
 from types import SimpleNamespace
 
-from sglang.srt.environ import temp_set_env
+from sglang.srt.utils.common import temp_set_env
 from sglang.srt.utils import kill_process_tree
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
